+++ tempodb/encoding/vparquet4/schema.go
@@ -46,14 +46,15 @@ const (
 
 // These definition levels match the schema below
 const (
-	DefinitionLevelTrace                     = 0
-	DefinitionLevelServiceStats              = 1
-	DefinitionLevelResourceSpans             = 1
-	DefinitionLevelResourceAttrs             = 2
-	DefinitionLevelResourceSpansILSSpan      = 3
-	DefinitionLevelResourceSpansILSSpanAttrs = 4
-	DefinitionLevelResourceSpansILSSpanEvent = 4
-	DefinitionLevelResourceSpansILSSpanLink  = 4
+	DefinitionLevelTrace                          = 0
+	DefinitionLevelServiceStats                   = 1
+	DefinitionLevelResourceSpans                  = 1
+	DefinitionLevelResourceAttrs                  = 2
+	DefinitionLevelResourceSpansILSSpan           = 3
+	DefinitionLevelResourceSpansILSSpanAttrs      = 4
+	DefinitionLevelResourceSpansILSSpanEvent      = 4
+	DefinitionLevelResourceSpansILSSpanLink       = 4
+	DefinitionLevelResourceSpansILSSpanEventAttrs = 5
 
 	FieldResourceAttrKey       = "rs.list.element.Resource.Attrs.list.element.Key"
 	FieldResourceAttrVal       = "rs.list.element.Resource.Attrs.list.element.Value.list.element"